refactor: reorganiza estrutura de pastas e padroniza nomenclatura de arquivos
