+++ src/routes/index.js
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import complaintRoutes from "./complaints.routes.js";
-import "../shared/utils/zod.config.js";
+import "../config/zod.config.js";
 
 const router = Router();
