+++ src/shared/errors/internal-server.error.js
@@ -1,5 +1,5 @@
-import { AppError } from "./appError.js";
-import { ERROR_CODES } from "./errorCodes.js";
+import { AppError } from "./app.error.js";
+import { ERROR_CODES } from "../types/error.codes.js";
 
 export class InternalServerError extends AppError {
   constructor(err) {
